+++ app/assets/javascripts/components/admin/animes/admin_animes_table.jsx
